test : LottoScoreBoard 테스트 추가 (기존 WinningBoardTest에서 네이밍 변경)
